+++ src/node2json_str.js
@@ -2,7 +2,7 @@
 
 const util = require("./util");
 const buildOptions = require("./util").buildOptions;
-const x2j = require("./x2j");
+const x2j = require("./xmlstr2xmlnode");
 
 //TODO: do it later
 const convertToJsonString = function(node, options) {
