+++ src/parser.js
@@ -1,20 +1,20 @@
 "use strict";
 
-const nodeToJson = require("./n2j");
-const xmlToNodeobj = require("./x2j");
-const x2j = require("./x2j");
+const nodeToJson = require("./node2json");
+const xmlToNodeobj = require("./xmlstr2xmlnode");
+const x2xmlnode = require("./xmlstr2xmlnode");
 const buildOptions = require("./util").buildOptions;
 
 exports.parse = function(xmlData, options) {
-    options = buildOptions(options,x2j.defaultOptions,x2j.props);
+    options = buildOptions(options, x2xmlnode.defaultOptions, x2xmlnode.props);
     return nodeToJson.convertToJson(xmlToNodeobj.getTraversalObj(xmlData, options), options);
 };
 exports.convertTonimn = require("../src/nimndata").convert2nimn;
 exports.getTraversalObj = xmlToNodeobj.getTraversalObj;
 exports.convertToJson = nodeToJson.convertToJson;
-exports.convertToJsonString = require("./n2j_str").convertToJsonString;
+exports.convertToJsonString = require("./node2json_str").convertToJsonString;
 exports.validate = require("./validator").validate;
-exports.j2xParser = require("./j2x");
+exports.j2xParser = require("./json2xml");
 exports.parseToNimn = function (xmlData,schema,options){
     return exports.convertTonimn(exports.getTraversalObj(xmlData,options), schema, options);
 };
