+++ src/xmlstr2xmlnode.js
